Initial upload
I just dragged the whole Assets folder into here. I'm still figuring this out.

I'm not sure if all of these .meta files are necessary, but I'll exclude them in the future if they're not. 

This is just my first commit. When I feel like I'm ready to start a serious project, I'll make a new one that's more tidy.